+++ public-app/src/app/details/[id]/page.tsx
@@ -13,6 +13,8 @@ import { toast } from "@/hooks/useToast";
 
 import { AccountEntity, PriceList, UploadResponse } from "@/types/account.type";
 
+import { getRankImageUrl } from "@/lib/utils";
+
 import { isAxiosError } from "axios";
 import Image from "next/image";
 import { notFound, useParams, useRouter } from "next/navigation";
@@ -178,22 +180,6 @@ export default function AccountDetailPage() {
     account.thumbnail && account.otherImages?.length
       ? [{ ...account.thumbnail, isThumbnail: true }, ...account.otherImages]
       : (account.otherImages ?? [account.thumbnail]);
-
-  function getRankImageUrl(rank: string): string {
-    const r = rank.toLowerCase();
-
-    if (r.includes("iron")) return "/rank/Iron 3.svg";
-    if (r.includes("bronze")) return "/rank/Bronze 3.svg";
-    if (r.includes("silver")) return "/rank/Silver 3.svg";
-    if (r.includes("gold")) return "/rank/Gold 3.svg";
-    if (r.includes("platinum")) return "/rank/Platinum 3.svg";
-    if (r.includes("diamond")) return "/rank/Diamond 3.svg";
-    if (r.includes("ascendant")) return "/rank/Ascendant 3.svg";
-    if (r.includes("immortal")) return "/rank/Immortal 3.svg";
-    if (r.includes("radiant")) return "/rank/Radiant 3.svg";
-
-    return "/ranks/Unranked.svg";
-  }
 
   return (
     <main className="min-h-screen bg-black text-white">
