+++ public-app/src/components/bookings/BookingDetail.tsx
@@ -1,5 +1,7 @@
 import { BookingWithAccountEntity } from "@/types/booking.type";
 
+import { getRankImageUrl } from "@/lib/utils";
+
 import { Instrument_Sans, Staatliches } from "next/font/google";
 import Image from "next/image";
 
@@ -20,25 +22,9 @@ type BookingDetailProps = {
 };
 
 export default function BookingDetail({ booking }: BookingDetailProps) {
-  function formatDuration(duration: string): string {
+  const formatDuration = (duration: string): string => {
     return duration.replace(/d/g, " Day(s)").replace(/h/g, " Hour(s)");
-  }
-
-  function getRankImageUrl(rank: string): string {
-    const r = rank.toLowerCase();
-
-    if (r.includes("iron")) return "/rank/Iron 3.svg";
-    if (r.includes("bronze")) return "/rank/Bronze 3.svg";
-    if (r.includes("silver")) return "/rank/Silver 3.svg";
-    if (r.includes("gold")) return "/rank/Gold 3.svg";
-    if (r.includes("platinum")) return "/rank/Platinum 3.svg";
-    if (r.includes("diamond")) return "/rank/Diamond 3.svg";
-    if (r.includes("ascendant")) return "/rank/Ascendant 3.svg";
-    if (r.includes("immortal")) return "/rank/Immortal 3.svg";
-    if (r.includes("radiant")) return "/rank/Radiant 3.svg";
-
-    return "/ranks/Unranked.svg";
-  }
+  };
 
   return (
     <div>
